+++ kipart/pckg_info.py
@@ -1,3 +1,3 @@
-__version__ = "0.1.35"
+__version__ = "0.1.36"
 __author__ = "XESS Corp."
 __email__ = "info@xess.com"