+++ kipart/xilinx7_reader.py
@@ -33,8 +33,13 @@ from .kipart import *
 defaulted_names = set(list())
 
 
-def xilinx7_reader(csv_file):
+def xilinx7_reader(part_data_file, part_data_file_name, part_data_file_type=".csv"):
     """Extract the pin data from a Xilinx CSV file and return a dictionary of pin data."""
+
+    # If part data file is Excel, convert it to CSV.
+    if part_data_file_type == ".xlsx":
+        part_data_file = convert_xlsx_to_csv(part_data_file)
+    csv_file = part_data_file
 
     # Create a dictionary that uses the unit numbers as keys. Each entry in this dictionary
     # contains another dictionary that uses the side of the symbol as a key. Each entry in