+++ HISTORY.rst
@@ -3,12 +3,18 @@
 History
 -------
 
+0.1.36 (2019-10-3119)
+______________________
+
+* KiPart now accepts part data stored in Excel .xlsx files.
+* Added reader for GOWIN FPGA pin tables.
+
+
 0.1.35 (2019-09-19)
 ______________________
 
 * Kipart now creates individual .lib files if given multiple .csv files with no global output .lib file specified using the -o option.
 * kilib2csv now creates individual .csv files if given multiple .lib files with no global output .csv file specified using the -o option.
-
 
 
 0.1.34 (2019-06-27)